refactor InMemoryCustomRefundStorage to leverage InMemoryStorage as a field
