+++ src/tests/run_manually.rs
@@ -24,8 +24,7 @@ use crate::zkevm_circuits::base_structures::vm_state::GlobalContextWitness;
 use crate::zkevm_circuits::main_vm::main_vm_entry_point;
 use circuit_definitions::aux_definitions::witness_oracle::VmWitnessOracle;
 use circuit_definitions::zk_evm::vm_state::cycle;
-use utils::storage::InMemoryCustomRefundStorage;
-use utils::StorageRefund;
+use storage::{InMemoryCustomRefundStorage, StorageRefund};
 use zkevm_assembly::Assembly;
 
 #[test]
@@ -261,7 +260,7 @@ pub(crate) fn run_with_options(entry_point_bytecode: Vec<[u8; 32]>, options: Opt
     let mut known_contracts = HashMap::new();
     known_contracts.extend(options.other_contracts.iter().cloned());
 
-    save_predeployed_contracts(&mut storage_impl, &mut tree, &known_contracts);
+    save_predeployed_contracts(&mut storage_impl.storage, &mut tree, &known_contracts);
 
     let mut basic_block_circuits = vec![];
