+++ src/tests/utils/mod.rs
@@ -1,10 +1,2 @@
 pub mod preprocess_asm;
-pub mod storage;
 pub mod testing_tracer;
-
-/// Enum holding the types of storage refunds
-#[derive(Debug, Copy, Clone)]
-pub(crate) enum StorageRefund {
-    Cold,
-    Warm,
-}
